moved common function into validators
